requirements.txt
datasets
kaggle
pandas
pymongo
python-dotenv
tqdm
rapidfuzz
langdetect
